Fix for multiple status_set() in assess_status()
This fixes a multiple status_set() bug in the implementation of assess_status()
on the charm, when it is determining whether it is active AND clustered.
The change hooks into the _determine_os_workload_status() directly in
charmhelpers to ensure that status_set() is only called once when
assess_status() is called.

Change-Id: Idfd93126c3edb41f83897c82ee7f20fe5f366559
Related-Bug:#1588462