Ensure rabbit hosts are resolvable
The update_hosts_file function requires a full map of entries to
write into /etc/hosts; the previous use of this function tried
to write individual host entries only.

Ensure that a full map of self + peers is written at appropriate
points in a units lifecycle to ensure self and peer hostname
resolvability.

(also fixed a typo in a status message)

Change-Id: If40b31a55641861a1e03985f318e012b545fcc25
Closes-Bug: 1592200
Closes-Bug: 1590085