Revert changes relating to management of /etc/hosts
Switching to active management by the charm of /etc/hosts
resulted in a number of race conditions during cluster
formation as a nasty side effect.

Revert changes to the charm related to this switch, as charm
versions prior to these changes are far more reliable to
deploy.

Further work is required to support MAAS 2.0 with Juju 2.0.

This reverts commit 0ded449e8d32428baf7992cfc6999f82e9996c1a.
This reverts commit 7ffdcd204e6550cf5dcad57ed07ea302c1eedcce.

Closes-Bug: 1592200

Change-Id: Ib2d6714ac1c4b32542cea02c24f75d7ffbde8282